projects/phase-08-finetune-platform/requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
trl>=0.7.0
wandb>=0.16.0
pyyaml>=6.0.0
rich>=13.0.0
